Move MapWindow to separate files
Separate MapWindow from bunch of classes in layouts.cpp/hpp
Inherit it from WindowPinnableBase.

TODO: map arrow (player position) updates orientation only
when you open the inventory. When the map window is pinned,
the arrow direction is not updated.
