+++ apps/openmw/mwgui/layouts.hpp
@@ -82,30 +82,6 @@ namespace MWGui
     MyGUI::TextBox* batchcounter;
   };
 
-  class MapWindow : public OEngine::GUI::Layout, public LocalMapBase
-  {
-  public:
-    MapWindow();
-    virtual ~MapWindow(){}
-
-    void setVisible(bool b);
-    void setPlayerPos(const float x, const float y);
-    void setPlayerDir(const float x, const float y);
-    void setCellName(const std::string& cellName);
-  
-  private:
-    void onDragStart(MyGUI::Widget* _sender, int _left, int _top, MyGUI::MouseButton _id);
-    void onMouseDrag(MyGUI::Widget* _sender, int _left, int _top, MyGUI::MouseButton _id);
-    void onWorldButtonClicked(MyGUI::Widget* _sender);
-
-    MyGUI::ScrollView* mGlobalMap;
-    MyGUI::ImageBox* mPlayerArrow;
-    MyGUI::Button* mButton;
-    MyGUI::IntPoint mLastDragPos;
-    bool mVisible;
-    bool mGlobal;
-  };
-
   class MainMenu : public OEngine::GUI::Layout
   {
   public:
